+++ CHANGELOG.md
@@ -4,11 +4,15 @@
 
 - [ ] add contact section (emailJs)
 - [ ] add projects languages types
-- [ ] find suitable BG to hero section
 - [ ] add [Qualifications](https://codepen.io/xichen/pen/wzpeMM) section Education & Work Experience as timeline (date - title - info)
 - [ ] apply [Alphabet soup](https://codepen.io/mohmmedtaysser/pen/rNQYVwg?editors=1010) effect on hero section
 
 ## History
+
+### 1 February 2025
+
+- Add hero section background image
+- Add "hand writing" effect on hero section
 
 ### 25 January 2025